Fixed first launch currency and language selection
