+++ app/src/main/java/com/stonks/ui/currency/CurrencyFragment.kt
@@ -1,6 +1,5 @@
 package com.stonks.ui.currency
 
-import android.content.res.Configuration
 import android.content.res.Resources
 import android.os.Bundle
 import android.util.Log
@@ -272,13 +271,5 @@ class CurrencyFragment(bottomNavigationHeight: Int) : Fragment() {
         params.height = Resources.getSystem().displayMetrics.heightPixels - (Resources.getSystem().displayMetrics.heightPixels - totalHeight) / 3
         currency_chart.requestLayout()
         currency_scroll_view.fullScroll(View.FOCUS_DOWN)
-    }
-
-    private fun updateLocale(languageCode: String) {
-        val locale = Locale(languageCode)
-        Locale.setDefault(locale)
-        val config: Configuration = resources.configuration
-        config.setLocale(locale)
-        resources.updateConfiguration(config, resources.displayMetrics)
     }
 }
